mtk: wifi: mt76: mt7996: fix incorrect interpretation of EHT MCS and NSS capabilities
The EHT-MCS Map (20 MHz-Only Non-AP STA) subfield of the Supported EHT-MCS And NSS Set field in the EHT Capabilities element is not present for AP. Therefore, STA should not parse the subfield.
Moreover, AP should parse the subfield only if STA is 20 MHz-Only, which is confirmed by checking the Supported Channel Width Set subfield in the HE Capabilities element.

Signed-off-by: Benjamin Lin <benjamin-jw.lin@mediatek.com>